appended dtbs works!
With this commit, the user can build a single binary with the zImage
linked in and the dtbs appended.

The user needs to provide a board-*.c file, address of the uart, then:

  make MFG=boardname APPEND_KERNEL=/tmp/zImage APPEND_DTBS=/tmp/*.dtb \
	uImage

Look at board-raumfeld.c for an example of how to determine the board
you are on and set the desired compatible string.

Signed-off-by: Jason Cooper <jason@lakedaemon.net>