boards: add Broadcom BCM63xx board code support
Read the CHIP_ID value from the INT_PERIPH_ID register and match a
generic "brcm,bcm63138" entry with DTBs that have been built-into the
impedance matcher.

In the future, we will probably need to read the boot strap to figure
out the boot device and read the board id from the NVRAM and match it
exactly with a collection of DTBs.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>