main: make kexec work
For kexec to work, we need to initialize the stack pointer before branching
to any function or pushing things to the stack.

First, make the function naked so gcc omits the prelude.

Also set SP to the location we were called from. This is possibly a bad
idea, so if it hits us we need to find a better place for the stack. On PXA
it works fine though.