libfdt/Makefile: Add libfdt.a target
libfdt.a was being created as a side-effect of compiling fdt_wip.o. This
starts to break down when building in parallel as the job to produce
fdt_wip.o (and therefore libfdt.a) may be launched before the job to
produce the other .o files has finished.

Add an explicit libfdt.a target so that make's dependency tracking can
correctly launch the compile/archive processes.

Signed-off-by: Chris Packham <judge.packham@gmail.com>